+++ src/Contests 15/ Simple Counter/counter.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.20;
 contract CommunityCounter {
     mapping(uint256 => uint256) totalCounterperAnnouncement;
 
-    function incrementCounter(uint announcement_id) public {
+    function incrementCounter(uint256 announcement_id) public {
         require(announcement_id > 0, "Not Valid ID");
         uint256 count = 1;
         totalCounterperAnnouncement[announcement_id] += count;
     }
 
-    function viewCounter(uint announcement_id) public view returns (uint) {
+    function viewCounter(uint256 announcement_id) public view returns (uint256) {
         return totalCounterperAnnouncement[announcement_id];
     }
 }
